admin/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
bcrypt>=4.1.2
pyjwt>=2.8.0
gunicorn>=21.2.0
redis>=5.0.0
celery>=5.3.0
pyzmq>=25.1.0
sqlalchemy>=2.0.0
